+++ Kernel/Modules/AdminUser.pm
@@ -40,6 +40,9 @@ sub Run {
     my $Search       = $ParamObject->GetParam( Param => 'Search' )       || '';
     my $Notification = $ParamObject->GetParam( Param => 'Notification' ) || '';
 
+    # Get list of valid IDs.
+    my @ValidIDList = $Kernel::OM->Get('Kernel::System::Valid')->ValidIDsGet();
+
     # ------------------------------------------------------------ #
     #  switch to user
     # ------------------------------------------------------------ #
@@ -120,6 +123,7 @@ sub Run {
             NoOutOfOffice => 1,
         );
         my $Output = $LayoutObject->Header();
+        $Output .= $Self->_ActiveAgentAccountLimitNote();
         $Output .= $LayoutObject->NavigationBar();
         $Output .= $LayoutObject->Notify( Info => Translatable('Agent updated!') )
             if ( $Notification && $Notification eq 'Update' );
@@ -181,8 +185,29 @@ sub Run {
             $Errors{'UserLoginInvalid'} = 'ServerError';
         }
 
+        # error if activating non-root agent account over limit
+        my $ActiveAgentAccountLimit = $ConfigObject->Get('ActiveAgentAccountLimit');
+        if ($ActiveAgentAccountLimit && ((!$GetParam{UserID}) || ($GetParam{UserID} != 1))) {
+            my $Valid = scalar grep { $_ eq $GetParam{ValidID} } @ValidIDList;
+            if ( $Valid ) {
+                my %User = $UserObject->GetUserData(
+                    UserID => $GetParam{UserID},
+                    Valid => 1,
+                );
+                if (!%User) {
+                    my $ActiveAgentNumber = $UserObject->ActiveUserNumber();
+                    if ($ActiveAgentNumber >= $ActiveAgentAccountLimit) {
+                        $Note .= $LayoutObject->Notify(
+                            Info => 'Operation failed - active agent limit exceeded!',
+                            Priority => 'Error',
+                        );
+                    }
+                }
+            }
+        }
+
         # if no errors occurred
-        if ( !%Errors )
+        if ( (!%Errors) && ($Note eq '') )
         {
 
             # update user
@@ -268,6 +293,7 @@ sub Run {
             }
         }
         my $Output = $LayoutObject->Header();
+        $Output .= $Self->_ActiveAgentAccountLimitNote();
         $Output .= $Note;
         $Output .= $LayoutObject->NavigationBar();
         $Self->_Edit(
@@ -295,6 +321,7 @@ sub Run {
         $GetParam{UserLogin} = $ParamObject->GetParam( Param => 'UserLogin' );
 
         my $Output = $LayoutObject->Header();
+        $Output .= $Self->_ActiveAgentAccountLimitNote();
         $Output .= $LayoutObject->NavigationBar();
         $Self->_Edit(
             Action => 'Add',
@@ -350,8 +377,20 @@ sub Run {
             $Errors{'UserLoginInvalid'} = 'ServerError';
         }
 
+        # error if creating valid, non-root agent account over limit
+        my $ActiveAgentAccountLimit = $ConfigObject->Get('ActiveAgentAccountLimit');
+        if ($ActiveAgentAccountLimit && ((!$GetParam{UserID}) || ($GetParam{UserID} != 1))) {
+            my $Valid = scalar grep { $_ eq $GetParam{ValidID} } @ValidIDList;
+            if ( $Valid ) {
+                my $ActiveAgentNumber = $UserObject->ActiveUserNumber();
+                if ($ActiveAgentNumber >= $ActiveAgentAccountLimit) {
+                    $Note .= 'Operation failed - active agent limit exceeded!';
+                }
+            }
+        }
+
         # if no errors occurred
-        if ( !%Errors )
+        if ( (!%Errors) && ($Note eq '') )
         {
 
             # add user
@@ -439,6 +478,7 @@ sub Run {
             }
         }
         my $Output = $LayoutObject->Header();
+        $Output .= $Self->_ActiveAgentAccountLimitNote();
         $Output .= $Note
             ? $LayoutObject->Notify(
             Priority => 'Error',
@@ -467,6 +507,7 @@ sub Run {
     else {
         $Self->_Overview( Search => $Search );
         my $Output = $LayoutObject->Header();
+        $Output .= $Self->_ActiveAgentAccountLimitNote();
         $Output .= $LayoutObject->NavigationBar();
         $Output .= $LayoutObject->Notify( Info => Translatable('Agent updated!') )
             if ( $Notification && $Notification eq 'Update' );
@@ -691,6 +732,40 @@ sub _Overview {
     }
 
     return 1;
+}
+
+sub _ActiveAgentAccountLimitNote {
+    my ( $Self, %Param ) = @_;
+
+    # get layout object
+    my $LayoutObject = $Kernel::OM->Get('Kernel::Output::HTML::Layout');
+
+    # get user object
+    my $UserObject = $Kernel::OM->Get('Kernel::System::User');
+
+    # get config object
+    my $ConfigObject = $Kernel::OM->Get('Kernel::Config');
+
+    # add warning if agent account limit reached
+    my $ActiveAgentAccountLimit = $ConfigObject->Get('ActiveAgentAccountLimit');
+    if ($ActiveAgentAccountLimit) {
+        my $ActiveAgentNumber = $UserObject->ActiveUserNumber();
+
+        if ($ActiveAgentNumber >= $ActiveAgentAccountLimit) {
+            my $ActiveAgentNumberLimitMessage = $ConfigObject->Get('ActiveAgentAccountLimitMessage')
+                || 'Active agents limit (%s) reached. No more active agent account allowed.';
+
+            $ActiveAgentNumberLimitMessage =
+                $LayoutObject->{LanguageObject}->Translate($ActiveAgentNumberLimitMessage, $ActiveAgentAccountLimit),
+
+            return $LayoutObject->Notify(
+                Priority => 'Warning',
+                Info => $ActiveAgentNumberLimitMessage,
+            );
+        }
+    }
+
+    return '';
 }
 
 1;